refactor(formatter_core,formatter,formatter_json): share `Expand` and `BracketSpacing` options (#22821)
We chose `core` as the location for duplicate `struct`s, but the final options will continue to be built by each formatter.
