+++ crates/oxc_formatter_json/src/lib.rs
@@ -21,5 +21,5 @@ mod separated;
 pub use crate::{
     context::JsonFormatContext,
     format::format,
-    options::{Expand, JsonFormatOptions, JsonVariant},
+    options::{JsonFormatOptions, JsonVariant},
 };
